requirements.txt
streamlit
requests
pydantic
pillow
numpy
moviepy
python-dotenv